WaterMeterNumberRecognition_FlyAI/requirements.txt
flyai
numpy
opencv-python
matplotlib
scipy
Pillow==6.2.2
torch==1.6.0
torchvision==0.7.0
timm
efficientnet_pytorch